Projects/linkedin_post_generator/requirements.txt
anthropic>=0.40.0
httpx>=0.27.0,<0.28.0
python-dotenv>=1.0.0
